+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/config/ClusterConfiguration.java
@@ -21,26 +21,18 @@ import org.apache.shenyu.admin.config.properties.ClusterProperties;
 import org.apache.shenyu.admin.config.properties.ClusterZookeeperProperties;
 import org.apache.shenyu.admin.mode.ShenyuRunningModeService;
 import org.apache.shenyu.admin.mode.cluster.filter.ClusterForwardFilter;
-import org.apache.shenyu.admin.mode.cluster.impl.jdbc.ClusterSelectMasterServiceJdbcImpl;
-import org.apache.shenyu.admin.mode.cluster.impl.zookeeper.ClusterSelectMasterServiceZookeeperImpl;
-import org.apache.shenyu.admin.mode.cluster.impl.zookeeper.ClusterZookeeperClient;
-import org.apache.shenyu.admin.mode.cluster.impl.zookeeper.ClusterZookeeperConfig;
-import org.apache.shenyu.admin.mode.cluster.mapper.ClusterMasterMapper;
 import org.apache.shenyu.admin.mode.cluster.service.ClusterSelectMasterService;
 import org.apache.shenyu.admin.mode.cluster.service.ShenyuClusterService;
 import org.apache.shenyu.admin.service.impl.UpstreamCheckService;
 import org.apache.shenyu.admin.service.manager.LoadServiceDocEntry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.integration.jdbc.lock.JdbcLockRegistry;
-import org.springframework.integration.zookeeper.lock.ZookeeperLockRegistry;
-
-import java.util.Objects;
 
 /**
  * The type Cluster configuration.
@@ -50,6 +42,23 @@ import java.util.Objects;
 public class ClusterConfiguration {
     
     private static final Logger LOGGER = LoggerFactory.getLogger(ClusterConfiguration.class);
+    
+    @Value("${server.servlet.context-path:}")
+    private String contextPath;
+    
+    @Value("${server.port:}")
+    private String port;
+    
+//    /**
+//     * Shenyu Admin distributed lock by spring-integration-zookeeper.
+//     *
+//     * @param lockRepository the lockRepository
+//     * @return the shenyu Admin register repository
+//     */
+//    @Bean
+//    public ZookeeperLockRegistry zookeeperLockRegistry(final LockRepository lockRepository) {
+//        return new ZookeeperLockRegistry(lockRepository);
+//    }
     
     /**
      * Shenyu running mode cluster service.
@@ -84,55 +93,6 @@ public class ClusterConfiguration {
     @ConditionalOnProperty(value = {"shenyu.cluster.enabled"}, havingValue = "true", matchIfMissing = false)
     public ClusterForwardFilter clusterForwardFilter() {
         return new ClusterForwardFilter();
-    }
-    
-    /**
-     * Shenyu select master service.
-     *
-     * @param clusterProperties the cluster properties
-     * @param jdbcLockRegistry the jdbc lock registry
-     * @param clusterMasterMapper the cluster master mapper
-     * @return the shenyu select master service
-     */
-    @Bean
-    @ConditionalOnProperty(value = {"shenyu.cluster.select-type"}, havingValue = "jdbc", matchIfMissing = true)
-    public ClusterSelectMasterService clusterSelectMasterJdbcService(final ClusterProperties clusterProperties,
-                                                                     final JdbcLockRegistry jdbcLockRegistry,
-                                                                     final ClusterMasterMapper clusterMasterMapper) {
-        return new ClusterSelectMasterServiceJdbcImpl(clusterProperties, jdbcLockRegistry, clusterMasterMapper);
-    }
-    
-    /**
-     * Shenyu cluster select master service.
-     *
-     * @param clusterProperties the cluster properties
-     * @param zookeeperLockRegistry the zookeeper lock registry
-     * @return the shenyu cluster select master service
-     */
-    @Bean
-    @ConditionalOnProperty(value = {"shenyu.cluster.select-type"}, havingValue = "zookeeper", matchIfMissing = false)
-    public ClusterSelectMasterService clusterSelectMasterZookeeperService(final ClusterProperties clusterProperties,
-                                                                          final ZookeeperLockRegistry zookeeperLockRegistry) {
-        return new ClusterSelectMasterServiceZookeeperImpl(clusterProperties, zookeeperLockRegistry);
-    }
-    
-    /**
-     * register zkClient in spring ioc.
-     *
-     * @param clusterZookeeperProperties the zookeeper configuration
-     * @return ClusterZookeeperClient {@linkplain ClusterZookeeperClient}
-     */
-    @Bean
-    @ConditionalOnProperty(value = {"shenyu.cluster.select-type"}, havingValue = "zookeeper", matchIfMissing = false)
-    public ClusterZookeeperClient clusterZookeeperClient(final ClusterZookeeperProperties clusterZookeeperProperties) {
-        int sessionTimeout = Objects.isNull(clusterZookeeperProperties.getSessionTimeout()) ? 3000 : clusterZookeeperProperties.getSessionTimeout();
-        int connectionTimeout = Objects.isNull(clusterZookeeperProperties.getConnectionTimeout()) ? 3000 : clusterZookeeperProperties.getConnectionTimeout();
-        ClusterZookeeperConfig zkConfig = new ClusterZookeeperConfig(clusterZookeeperProperties.getUrl());
-        zkConfig.setSessionTimeoutMilliseconds(sessionTimeout)
-                .setConnectionTimeoutMilliseconds(connectionTimeout);
-        ClusterZookeeperClient client = new ClusterZookeeperClient(zkConfig);
-        client.start();
-        return client;
     }
     
 }
