+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/config/ClusterZookeeperConfiguration.java
@@ -1,0 +1,86 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shenyu.admin.config;
+
+import org.apache.shenyu.admin.config.properties.ClusterProperties;
+import org.apache.shenyu.admin.config.properties.ClusterZookeeperProperties;
+import org.apache.shenyu.admin.mode.cluster.impl.zookeeper.ClusterSelectMasterServiceZookeeperImpl;
+import org.apache.shenyu.admin.mode.cluster.impl.zookeeper.ClusterZookeeperClient;
+import org.apache.shenyu.admin.mode.cluster.impl.zookeeper.ClusterZookeeperConfig;
+import org.apache.shenyu.admin.mode.cluster.service.ClusterSelectMasterService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.integration.zookeeper.lock.ZookeeperLockRegistry;
+
+import java.util.Objects;
+
+/**
+ * The type Cluster zookeeper configuration.
+ */
+@Configuration(proxyBeanMethods = false)
+@EnableConfigurationProperties({ClusterProperties.class, ClusterZookeeperProperties.class})
+@ConditionalOnProperty(value = {"shenyu.cluster.select-type"}, havingValue = "zookeeper", matchIfMissing = false)
+public class ClusterZookeeperConfiguration {
+    
+    private static final Logger LOGGER = LoggerFactory.getLogger(ClusterZookeeperConfiguration.class);
+    
+    @Value("${server.servlet.context-path:}")
+    private String contextPath;
+    
+    @Value("${server.port:}")
+    private String port;
+    
+    /**
+     * Shenyu cluster select master service.
+     *
+     * @param clusterProperties the cluster properties
+     * @param zookeeperLockRegistry the zookeeper lock registry
+     * @return the shenyu cluster select master service
+     */
+    @Bean
+    @ConditionalOnProperty(value = {"shenyu.cluster.select-type"}, havingValue = "zookeeper", matchIfMissing = false)
+    public ClusterSelectMasterService clusterSelectMasterZookeeperService(final ClusterProperties clusterProperties,
+                                                                          final ZookeeperLockRegistry zookeeperLockRegistry) {
+        return new ClusterSelectMasterServiceZookeeperImpl(clusterProperties, zookeeperLockRegistry);
+    }
+    
+    /**
+     * register zkClient in spring ioc.
+     *
+     * @param clusterZookeeperProperties the zookeeper configuration
+     * @return ClusterZookeeperClient {@linkplain ClusterZookeeperClient}
+     */
+    @Bean
+    @ConditionalOnProperty(value = {"shenyu.cluster.select-type"}, havingValue = "zookeeper", matchIfMissing = false)
+    public ClusterZookeeperClient clusterZookeeperClient(final ClusterZookeeperProperties clusterZookeeperProperties) {
+        int sessionTimeout = Objects.isNull(clusterZookeeperProperties.getSessionTimeout()) ? 3000 : clusterZookeeperProperties.getSessionTimeout();
+        int connectionTimeout = Objects.isNull(clusterZookeeperProperties.getConnectionTimeout()) ? 3000 : clusterZookeeperProperties.getConnectionTimeout();
+        ClusterZookeeperConfig zkConfig = new ClusterZookeeperConfig(clusterZookeeperProperties.getUrl());
+        zkConfig.setSessionTimeoutMilliseconds(sessionTimeout)
+                .setConnectionTimeoutMilliseconds(connectionTimeout);
+        ClusterZookeeperClient client = new ClusterZookeeperClient(zkConfig);
+        client.start();
+        return client;
+    }
+    
+}
